Add the initial implementation of higher kinds
This is as close to the brand implementation as I can come thus far. The
situation in Carp is slightly different from that of OCaml, of course,
but we can leverage Carp's interfaces, macros, and module system in
combination to yield a fairly powerful result--uniquely defined brands,
projections, and injections.

I've tried to adopt terminology for types and functions that's easier to
comprehend and hopefully has less academic musk wafting around it.