Add kind? and mark-kind
Redefining kinds has proven problematic, at least in the REPL--it can
lead to scenarios that produce invalid c. Assuming the definition of a
given type doesn't change later in a program, (which it shouldn't as
this can also yield invalid c), it's safe to define the kind bindings
(brands, etc.) only once.

The two functions added in this commit make it possible to determine
whether or not kind bindings have already been generated for a type, and
thus to selectively generate them where needed.