Cleanup Functor
Minor rewrites and ergonomics improvements to Functor. The most notable
changes are:

- We explicitly check whether or not a type deriving fmap is a proper
kind given by Kinds.kind.
- We explicitly check whether or not a kind is an instance of functor,
i.e. whether or not it implements fmap.

The decision to make kind generation explicit on the part of the
programmer simplifies implementation on the library side--while this is
not as convenient for the programmer, it's only one more short line of
code, and it seems expedient considering the library itself is not yet
complete. We can add such conveniences back later in its lifetime.