Add applicative typeclass
This commit adds the applicative typeclass and a new FP module for
functional utility functions--mostly to permit currying.

Much of the Applicative typeclass's flexibility in Haskell's relies on
lazy evaluation and currying. It's a bit cumbersome to curry functions
in Carp at the moment, so it may be more advantageous to define variants
of sequence for multi-arity functions in the future.