Functor: Add "endo" variants of functor functions
In addition to the traditional, functional style functor definitions,
I've added support for 'in-place' variations, since Carp allows us to
manipulate structures in-place. Each in-place variant of a typeclass
function is suffixed with a `!`, following Carp convention.

This commit also adds the void function, which takes a functor
inhabitant to the Unit.