Functor: update docs and signatures; remove `void!`
The original signature of `fmap!` did not make much sense; since we're
changing values *in-place* the function passed to `fmap!` must be
`endomorphic`--that is, its argument and return types must be the same.
This is in keeping with the core library's `endo-map` functions over
arrays.

After the change to `fmap!`s signature, `void!` no longer makes much
sense, as one can only use it over a functor that already contained the
Unit type.