Applicative: Refactor definitions; remove FP
After a bit of laborious investigation, I discovered that
`FP.binary-partial` did not work to define applicative lifts because the
lack of a reference to the captured lambda variable leads to a double
free that isn't currently caught by the carp compiler.

This commit just inlines our lambdas instead, and fixes their signatures
to prevent the double free. Also added documentation.